Implementing the softmax and argmax layers. This branch is based on changes that are to be introduced in [PR 636](https://dev.azure.com/PerceptiLabs/PerceptiLabs/_git/PerceptiLabs/pullrequest/636), so the diff against master might look rather large for a while.

Changes:
* Added argmax implementation. All files in [backend/perceptilabs/layers/mathargmax/](https://dev.azure.com/PerceptiLabs/PerceptiLabs/_git/PerceptiLabs/pullrequest/650?_a=files&path=%2Fbackend%2Fperceptilabs%2Flayers%2Fmathargmax)
* Added softmax implementation. All files in [backend/perceptilabs/layers/softmax/](https://dev.azure.com/PerceptiLabs/PerceptiLabs/_git/PerceptiLabs/pullrequest/650?_a=files&path=%2Fbackend%2Fperceptilabs%2Flayers%2Fmathsoftmax)
* Added the new files to [included_files.txt](https://dev.azure.com/PerceptiLabs/PerceptiLabs/_git/PerceptiLabs/pullrequest/650?_a=files&path=%2Fbackend%2Fincluded_files.txt)
* Updated [package_data.json](https://dev.azure.com/PerceptiLabs/PerceptiLabs/_git/PerceptiLabs/pullrequest/650?_a=files&path=%2Fbackend%2Fpackage_data.json) to search for Jinja files in the subdirectories of perceptilabs/layer/
* Fixed an issue in [this file](https://dev.azure.com/PerceptiLabs/PerceptiLabs/_git/PerceptiLabs/pullrequest/650?_a=files&path=%2Ffrontend%2Fsrc%2Frenderer%2Fcomponents%2Fnetwork-elements%2Felements%2Fmath-argmax%2FviewBox-math-argmax.vue) where argmax data wasn't fetched correctly **[Frontend]**

Related work items: #1011

+++ backend/perceptilabs/coreInterface.py
@@ -868,7 +868,7 @@ class coreLogic():
 
         elif layerType in ["MathMerge", "MathSoftmax", "MathArgmax", "MathSwitch", "ProcessOneHot", "ProcessCrop", "ProcessReshape", "ProcessRescale"]:
             D=self.getStatistics({"layerId":layerId,"variable":"Y","innervariable":""})[-1]
-            output = createDataObject([np.squeeze(D)])
+            output = createDataObject([np.squeeze(D).astype(np.float32)])
             return {"Output":output}
         elif layerType == "ProcessGrayscale":
             D=self.getStatistics({"layerId":layerId,"variable":"Y","innervariable":""})[-1]
